Task 10: - A new library "Oresundbron" has been created.  There's references to the XML and DLL.
- The library has to classes "OresundCar", and "OresundMC". The baseprice for a car is 460 kr, with Brobizz it's 178 kr.
The baseprice for a MC is 235 kr, with Brobizz it's 92 kr.
The "VehicleType()"-method returns "Oresund car"/"Oresund MC".

- There's added 6 tests. 3 for the Oresund cars, and 3 for the Oresund MCs.
The first ones are to test the baseprice for both cars and MCs.
The next are to test the baseprice + Brobizz for both cars and MCs.
The last ones are to test the correct string return from the "VehicleType()"-method.

- The tests passed :)

- The code coverage is still good. 189 covered lines, and 2 not-covered lines.